Increase minimum charge
Electride search was finding too many electride sites at minimum charge of 0.1. This isn't a huge issue, but it found several points in Ti2C  that make understanding the results difficult. In the future I should probably provide a way to change this as a parameter in the input file. That way users would be able to adjust if they want more or fewer electride sites when modeling electrides similar to Ti2C.