Fix bug in seeded HSE workflows
HSE calculations seeded with PBE give an error. I think the WAVECAR grids weren't matching up so I'm switching to the same grid settings for both. I'm also adding a fix for the electride site finding algorithm so that two charge peaks that are basically on top of one another will be combined into one.